+++ Pyramid Scheme/CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to Pyramid Scheme will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.11]
 
-No unreleased changes at this time.
+### Added
+
+### Changed
+
+
+### Fixed
+- **Fixed mobile functionality**
+  - Adjusted mobile screen settings to fit 5 investors
 
 ---